Add a bit more regex

+++ hand.history/Models/Player.cs
@@ -6,12 +6,12 @@ namespace hand.history.Models
 {
     public class Player
     {
-        public decimal Id { get; }
+        public double Id { get; }
 
         public string Username { get; }
         public bool Alive { get; set; }
         public int Position { get; set; }
-        public decimal Stack { get; set; }
+        public double Stack { get; set; }
 
         public Hand Hand { get; }
     }
